Creation of the recipe website's architecture
Creation of the different directories, files and images required for the project.
The project consists of a main page (index.html) itself constituted of a title and three links.
Every link refers to a recipe.
Each recipe's page contains:
- A title
- An image
- A small description
- The list of ingredients
- And the list of steps
All of this represents the first step (part HTML) of this global project.